# HotelBookingDetails/HotelBookingDetails.WasmClient/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "HotelBookingDetails.WebApi",
    "version": "1.0"
  },
  "paths": {
    "/api/Client": {
      "get": {
        "tags": [
          "Client"
        ],
        "summary": "Запрос возвращающий список всех клиентов",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Client"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Client"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Client"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Client"
        ],
        "summary": "Запрос на добавления клиента",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClientDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Client/{id}": {
      "get": {
        "tags": [
          "Client"
        ],
        "summary": "Запрос клиента по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Client"
        ],
        "summary": "Запрос на изменение клиента по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClientDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Client"
        ],
        "summary": "Удаление клиента по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Hotel": {
      "get": {
        "tags": [
          "Hotel"
        ],
        "summary": "Запрос возвращающий список всех отелей",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Hotel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Hotel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Hotel"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Hotel"
        ],
        "summary": "Запрос на добавления отеля",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotelDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/HotelDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/HotelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Hotel/{id}": {
      "get": {
        "tags": [
          "Hotel"
        ],
        "summary": "Запрос отеля по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Hotel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hotel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hotel"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Hotel"
        ],
        "summary": "Запрос на изменение клиента по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotelDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/HotelDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/HotelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Hotel"
        ],
        "summary": "Удаление клиента по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Passport": {
      "get": {
        "tags": [
          "Passport"
        ],
        "summary": "Запрос возвращающий список всех паспортов",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Passport"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passport"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passport"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Passport"
        ],
        "summary": "Запрос на добавления паспорта",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PassportDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PassportDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PassportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Passport/{id}": {
      "get": {
        "tags": [
          "Passport"
        ],
        "summary": "Запрос паспорта по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Passport"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passport"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passport"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Passport"
        ],
        "summary": "Запрос на изменение паспорта по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PassportDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PassportDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PassportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Passport"
        ],
        "summary": "Удаление паспорта по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Client": {
        "required": [
          "birthday",
          "fullName",
          "id",
          "passportData"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "fullName": {
            "minLength": 1,
            "type": "string"
          },
          "passportData": {
            "$ref": "#/components/schemas/Passport"
          },
          "birthday": {
            "$ref": "#/components/schemas/DateOnly"
          }
        },
        "additionalProperties": false
      },
      "ClientDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "passportDataId": {
            "type": "integer",
            "format": "int32"
          },
          "birthdayYear": {
            "maximum": 2024,
            "minimum": 1900,
            "type": "integer",
            "format": "int32"
          },
          "birthdayMonth": {
            "maximum": 12,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "birthdayDay": {
            "maximum": 31,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "DateOnly": {
        "type": "object",
        "properties": {
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "month": {
            "type": "integer",
            "format": "int32"
          },
          "day": {
            "type": "integer",
            "format": "int32"
          },
          "dayOfWeek": {
            "$ref": "#/components/schemas/DayOfWeek"
          },
          "dayOfYear": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "dayNumber": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "DayOfWeek": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
      "Hotel": {
        "required": [
          "address",
          "city",
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "address": {
            "minLength": 1,
            "type": "string"
          },
          "city": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "HotelDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Passport": {
        "required": [
          "id",
          "number",
          "series"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "series": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "PassportDto": {
        "type": "object",
        "properties": {
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "series": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}